+++ source/CargoSpace/Assets/Scripts/Logging/LogManager.cs.meta
